VoteBot votes for emoji! :ballot_box_with_check:
Sadly, VoteBot has a sweet tooth and votes for :doughnut: 15% of the time.
Furthermore, VoteBot is juvenile and votes for :poop: 20% of the time.

When not voting for :doughnut: or :poop:, VoteBot can’t be bothered to
pick a favorite, so it picks one at random. C'mon VoteBot, try harder!