Normalize emojivoto k8s config for inject testing
emojivoto is our primary example for conduit inject, but running conduit
inject on this file causes minor reordering and addition of default
fields, making it harder to diff against the original.

Modify the emojivoto k8s config to be logically equivalent to the
existing version, but organized in a way that makes diff'ing against an
injected copy easier.

Signed-off-by: Andrew Seigner <siggy@buoyant.io>