Ensure that running telepresence daemons are stopped
The `setup_dev_env.sh` script sets `KUBECONFIG` to a relative path. This
means that it will be resolved by the Telepresence user daemon when it
connects to the cluster. If telepresence was started in some other
directory prior to running this example, then the user daemon will
resolve the path incorrectly and fail to connect, and the connector.log
contains the message:
```
unable to track k8s cluster: stat ./emojivoto_k8s_context.yaml: no such file or directory
```

This commit adds a `telepresence quit -ur` call prior to calling
`telepresence version`, thereby ensuring that the daemons are stopped.